+++ src/DebugCodeLens/DebugCodeLensProvider.ts
@@ -1,9 +1,9 @@
 import * as vscode from 'vscode'
 import { extensionName } from '../appGlobals'
+import { escapeRegExp } from '../helpers'
 import { basename } from 'path'
 import { DebugCodeLens } from './DebugCodeLens'
-import { TestReconciliationState } from '../TestReconciliationState'
-import { TestResultProvider } from '../TestResultProvider'
+import { TestReconciliationState, TestResultProvider } from '../TestResults'
 
 export class DebugCodeLensProvider implements vscode.CodeLensProvider {
   private _enabled: boolean
@@ -56,7 +56,7 @@ export class DebugCodeLensProvider implements vscode.CodeLensProvider {
   resolveCodeLens(codeLens: vscode.CodeLens, _: vscode.CancellationToken): vscode.ProviderResult<vscode.CodeLens> {
     if (codeLens instanceof DebugCodeLens) {
       codeLens.command = {
-        arguments: [codeLens.fileName, codeLens.testName],
+        arguments: [codeLens.fileName, escapeRegExp(codeLens.testName)],
         command: `${extensionName}.run-test`,
         title: 'Debug',
       }