+++ src/TestResults/TestResultProvider.ts
@@ -1,29 +1,8 @@
 import { TestReconciler, FormattedTestResults } from 'jest-editor-support'
 import { TestFileAssertionStatus } from 'jest-editor-support'
 import { TestReconciliationState } from './TestReconciliationState'
-import { parseTest } from './TestParser'
-import * as path from 'path'
-
-type Position = {
-  /** Zero-based column number */
-  column: number
-
-  /** Zero-based line number */
-  line: number
-}
-
-export type TestResult = {
-  name: string
-  start: Position
-  end: Position
-
-  status: TestReconciliationState
-  shortMessage?: string
-  terseMessage?: string
-
-  /** Zero-based line number */
-  lineNumberOfError?: number
-}
+import { TestResult } from './TestResult'
+import { parseTest } from '../TestParser'
 
 type TestResultsMap = { [filePath: string]: TestResult[] }
 
@@ -43,6 +22,10 @@ export class TestResultProvider {
 
   constructor() {
     this.reconciler = new TestReconciler()
+    this.resetCache()
+  }
+
+  resetCache() {
     this.resultsByFilePath = {}
     this.sortedResultsByFilePath = {}
   }
@@ -118,20 +101,7 @@ export class TestResultProvider {
   }
 
   updateTestResults(data: FormattedTestResults): TestFileAssertionStatus[] {
-    this.resultsByFilePath = {}
-    this.sortedResultsByFilePath = {}
-
-    // To support Windows systems, the drive letter is converted to a lowercase
-    // letter to match the convention of the document URI (e.g.: document.fileName)
-    if (data.testResults && path.sep === '\\') {
-      for (let i = 0; i < data.testResults.length; i += 1) {
-        if (data.testResults[i].name.match(/^[A-Z]:\\/)) {
-          const filePath = data.testResults[i].name
-          data.testResults[i].name = filePath[0].toLowerCase() + filePath.slice(1)
-        }
-      }
-    }
-
+    this.resetCache()
     return this.reconciler.updateFileWithJestStatus(data)
   }