+++ tests/TestResults/TestResultProvider.test.ts
@@ -1,4 +1,4 @@
-jest.unmock('../src/TestResultProvider')
+jest.unmock('../../src/TestResults/TestResultProvider')
 
 const updateFileWithJestStatus = jest.fn()
 const assertionsForTestFile = jest.fn()
@@ -30,9 +30,9 @@ jest.mock('path', () => {
   return path
 })
 
-import { TestResultProvider } from '../src/TestResultProvider'
-import { TestReconciliationState } from '../src/TestReconciliationState'
-import { parseTest } from '../src/TestParser'
+import { TestResultProvider } from '../../src/TestResults/TestResultProvider'
+import { TestReconciliationState } from '../../src/TestResults'
+import { parseTest } from '../../src/TestParser'
 
 describe('TestResultProvider', () => {
   describe('getResults()', () => {
@@ -149,28 +149,24 @@ describe('TestResultProvider', () => {
   })
 
   describe('updateTestResults()', () => {
-    afterEach(() => pathProperties.sep.mockReset())
-
-    it('should update the cached file status', () => {
+    it('should reset the cache', () => {
       const sut = new TestResultProvider()
       const results: any = {}
+      sut.resetCache = jest.fn()
       sut.updateTestResults(results)
 
-      expect(updateFileWithJestStatus).toBeCalledWith(results)
+      expect(sut.resetCache).toBeCalled()
     })
 
-    it('should test file paths to use lowercase drive letters on Windows', () => {
-      pathProperties.sep.mockReturnValue('\\')
+    it('should update the cached file status', () => {
+      const expected: any = {}
+      updateFileWithJestStatus.mockReturnValueOnce(expected)
 
       const sut = new TestResultProvider()
-      const results: any = {
-        testResults: [{ name: 'relative.js' }, { name: 'c:\\stays-lowercase' }, { name: 'D:\\changes-case' }],
-      }
-      sut.updateTestResults(results)
+      const results: any = {}
 
-      expect(updateFileWithJestStatus).toBeCalledWith({
-        testResults: [{ name: 'relative.js' }, { name: 'c:\\stays-lowercase' }, { name: 'd:\\changes-case' }],
-      })
+      expect(sut.updateTestResults(results)).toBe(expected)
+      expect(updateFileWithJestStatus).toBeCalledWith(results)
     })
   })
 })